Misc Cleanup (#138)
* Reject non-canonical IPv4 network addresses on subnet creation

* Bind all selected subnets in the Azure import wizard

* Fix bulk import name disambiguation looping on 50-character collisions

* Allow CIDR decrease for subnets with a grandparent or grandchildren

* Reset Azure reconcile confirmation when the review selection changes

* Truncate Azure VNet name to the subnet name column length on import

* Allow host IP assignments in /31 and /32 subnets

* Fix phantom unallocated range when allocations reach the top of the address space

* Validate parent eligibility before marking it fully allocated on Azure import

* Widen subnet name to 100 characters to fit Azure names

* Widen subnet description to 1000 characters and guard the fully-allocated note

* Re-lock the bulk import plan whenever the subscription selection changes

* Fix widen-CIDR address search for subnets at or above 128.0.0.0

* Give the subnet lock acquire a command timeout longer than its lock wait

* Log subnet lock release failures instead of surfacing them to callers

* Validate BASTET_FRAME_ANCESTORS at startup instead of failing every request

* Strip line breaks from values logged by the sanitization filter

* Show TempData messages on subnet details and fix the host IP child-subnet guard

* Remove the unused subnet division service

* Document the accepted reconcile check-and-act window

* Reject Azure resource IDs longer than the column instead of failing at insert

* Remove the orphaned error code plumbing from the error pages

* State both possible causes when reconcile cannot find a VNet

+++ src/Bastet/Services/IpUtilityService.cs
@@ -425,8 +425,10 @@ public class IpUtilityService : IIpUtilityService
             mergedRanges.Add(current);
         }
 
-        // Find gaps between allocated ranges
-        uint currentPosition = startIp;
+        // Find gaps between allocated ranges. This is a long because the position after the last
+        // address is one past the end of the IPv4 space: as a uint it wraps to 0, and the trailing
+        // check below then reads a fully allocated subnet as having the whole address space free.
+        long currentPosition = startIp;
 
         foreach ((uint Start, uint End) in mergedRanges)
         {
@@ -438,7 +440,7 @@ public class IpUtilityService : IIpUtilityService
                 {
                     unallocatedRanges.Add(new IPRange
                     {
-                        StartIp = UIntToIpString(currentPosition),
+                        StartIp = UIntToIpString((uint)currentPosition),
                         EndIp = UIntToIpString(Start - 1),
                         AddressCount = (Start - currentPosition > 2) ? Start - currentPosition - 1 : Start - currentPosition
                     });
@@ -447,7 +449,7 @@ public class IpUtilityService : IIpUtilityService
                 {
                     unallocatedRanges.Add(new IPRange
                     {
-                        StartIp = UIntToIpString(currentPosition),
+                        StartIp = UIntToIpString((uint)currentPosition),
                         EndIp = UIntToIpString(Start - 1),
                         AddressCount = Start - currentPosition
                     });
@@ -455,7 +457,7 @@ public class IpUtilityService : IIpUtilityService
             }
 
             // Move position to after this allocated range
-            currentPosition = End + 1;
+            currentPosition = (long)End + 1;
         }
 
         // Check if there's space after the last allocated range
@@ -476,7 +478,7 @@ public class IpUtilityService : IIpUtilityService
                 {
                     unallocatedRanges.Add(new IPRange
                     {
-                        StartIp = UIntToIpString(currentPosition),
+                        StartIp = UIntToIpString((uint)currentPosition),
                         EndIp = UIntToIpString(lastIp),
                         AddressCount = (lastIp - currentPosition > 0) ? lastIp - currentPosition : 1
                     });
@@ -485,7 +487,7 @@ public class IpUtilityService : IIpUtilityService
                 {
                     unallocatedRanges.Add(new IPRange
                     {
-                        StartIp = UIntToIpString(currentPosition),
+                        StartIp = UIntToIpString((uint)currentPosition),
                         EndIp = UIntToIpString(lastIp),
                         AddressCount = lastIp - currentPosition + 1
                     });
